Mech: Add player input, movement and collision
Adding player input was way easier than i initially thought.

Instead of relying on a timer to move the pieces. Learnt how to manipulate the _process function to achieve the result.

collision relied on get_cell_source_id() and checking if the place we want to move into is empty or not. If it is VOILA move otherwise dont.